refactor(fake-server): split FakeWaServer god class into focused modules
Breaks the 2334-line FakeWaServer into 5 cohesive files:

  FakeWaServer.ts       2334 → 1100 lines  (facade + lifecycle + media)
  iq-handlers.ts           — →  681 lines  (all 20 IQ handler registrations)
  ServerRegistries.ts      — →  522 lines  (peer/group/privacy/profile state)
  AppStateSyncManager.ts   — →  242 lines  (sync keys + mutation decrypt)
  PreKeyDispenser.ts       — →  183 lines  (prekey bundle + dispenser)

FakeWaServer remains the public facade — every existing public method
signature is preserved as a thin forwarding call to the appropriate
delegate. Tests, bench, and index.ts imports are unaffected.

Extracted modules:

- ServerRegistries: owns peerRegistry, groupRegistry, privacySettings,
  blocklistJids, profilePicturesByJid, businessProfilesByJid,
  latestStatusText, issuedPrivacyTokens, abPropsInput, all 8 listener
  fan-out sets, all onOutbound* subscriber methods, all *Snapshot
  accessors, createFakeGroup, lookupDeviceIdsForUser, and the
  toUserJidPart / toDeviceIdPart helpers.

- PreKeyDispenser: owns capturedPreKeyBundle, dispenser cursor +
  misses counter, preKeyBundleWaiters, triggerPreKeyUpload (with
  force option), awaitPreKeyBundle, dispenseOneTimePreKey.

- AppStateSyncManager: owns appStateCollectionProviders,
  appStateSyncKeysByKeyId, appStateCrypto, outboundMutationListeners,
  registerAppStateSyncKey, onOutboundAppStateMutation,
  expectAppStateMutation, provideAppStateCollection,
  pushServerSyncNotification, consumeOutboundAppStatePatches.

- iq-handlers.ts: exports registerDefaultIqHandlers(router, deps)
  which contains all ~20 router.register() calls that were inline in
  the constructor. Defines an IqHandlerDeps interface that receives
  narrow state slices from the 3 delegates — no circular deps.

Dependency graph (all arrows downward, no cycles):
  FakeWaServer → iq-handlers, ServerRegistries, PreKeyDispenser, AppStateSyncManager
  iq-handlers  → ServerRegistries (type-only via IqHandlerDeps)

144/144 tests pass, typecheck clean, lint 0 errors.
